+++ ascii/ascii_test.go
