+++ ascii/valid_pint_asm.go
