+++ ascii/valid_print_amd64.s
