+++ ascii/valid_print_default.go
